dev-requirements.txt
pytest
boto3>=1.26.0
PyYAML>=6.0
